restructuring for maven compiling
