+++ server/java_server/.gitignore
@@ -1,0 +1,5 @@
+# Maven target directory
+/target/
+
+# VS Code
+.vscode/
