+++ server/java_server/README.md
@@ -1,0 +1,4 @@
+Compiling and running: 
+
+mvn clean package
+java -jar target/java-server-1.0-SNAPSHOT.jar
